pyproject.toml
[tool.poetry]
name = "veriff-face-api"
version = "0.1.0"
description = ""
authors = ["kaspar <kasparkruus032@gmail.com>"]

packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
